backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djoser==2.3.1
webcolors==1.11.1
psycopg2-binary==2.9.11
pillow==11.0.0
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
PyYAML==6.0.1
environs==15.0.1
gunicorn==25.3.0
drf-extra-fields==3.7.0
django-filter==24.3